pyproject.toml
[project]
name = "opdroid"
version = "0.1.6"
description = "LLM-controlled Android device automation via ADB"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Shikhar Mishra" }
]
keywords = ["android", "adb", "llm", "automation", "ai-agent", "vision"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Testing",
    "Topic :: System :: Hardware :: Hardware Drivers",
]
dependencies = [
    "adbutils>=2.0.0",
    "litellm>=1.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pillow>=10.0.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/shikhr/opdroid"
Repository = "https://github.com/shikhr/opdroid"
Issues = "https://github.com/shikhr/opdroid/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/android_controller"]

[project.scripts]
opdroid = "android_controller.main:app"
